Major Refactor, added complete spider, items, item pipelines and cleaned code :fire:
