+++ scrapper/faculdades.txt
@@ -229,9 +229,7 @@
 Últimas Notícias da U.Porto</a></h3>
 		<ul class="latest-list">
 		
-<li><h5>12 Outubro   2017</h5><h4><a href="http://noticias.up.pt/universidade-do-porto-nas-100-melhores-do-mundo-em-cinco-areas-cientificas/" rel="nofollow" class="title">U.Porto nas 100 melhores do mundo em cinco áreas científicas</a></h4><p>Performance Ranking of Scientific Papers for World Universities 2017 coloca a U.Porto entre as 100 melhores universidades do mundo em cinco áreas de investigação científica.
-
- ... <a href="http://noticias.up.pt/universidade-do-porto-nas-100-melhores-do-mundo-em-cinco-areas-cientificas/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>12 Outubro   2017</h5><h4><a href="http://noticias.up.pt/u-porto-premiada-na-european-innovation-academy-2017/" rel="nofollow" class="title">U.Porto premiada na European Innovation Academy 2017</a></h4><p>Pedro Fonte, um dos selecionados pela U.Porto Inovação, com o apoio do Banco Santander, conduziu a sua equipa até ao pitch final, sagrando-se vencedor. ... <a href="http://noticias.up.pt/u-porto-premiada-na-european-innovation-academy-2017/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>11 Outubro   2017</h5><h4><a href="http://noticias.up.pt/o-uprunning-esta-de-volta/" rel="nofollow" class="title">O UPRunning está de volta</a></h4><p>O grupo de corrida da Universidade do Porto está de volta, todas as quartas-feiras, é gratuito e aberto a toda a comunidade! ... <a href="http://noticias.up.pt/o-uprunning-esta-de-volta/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>11 Outubro   2017</h5><h4><a href="http://noticias.up.pt/seniores-aprendem-segredos-da-alimentacao-saudavel-na-fcnaup/" rel="nofollow" class="title">Seniores aprendem segredos da alimentação saudável na FCNAUP</a></h4><p>Segunda edição do curso Culinária Saudável 65+ decorre a 11 de novembro e destina-se a pessoas com 65 ou mais anos e/ou aos seus cuidadores. Inscrições abertas. ... <a href="http://noticias.up.pt/seniores-aprendem-segredos-da-alimentacao-saudavel-na-fcnaup/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>11 Outubro   2017</h5><h4><a href="http://noticias.up.pt/literacia-na-gestao-da-diabetes-tipo-2-a-influencia-da-familia-e-dos-profissionais-de-saude/" rel="nofollow" class="title">Familiares e profissionais de saúde influenciam literacia de doentes crónicos</a></h4><p>Estudo do ISPUP demonstra que a literacia em saúde de pessoas com doenças crónicas como a diabetes tipo 2 é profundamente influenciada pela rede social do paciente. ... <a href="http://noticias.up.pt/literacia-na-gestao-da-diabetes-tipo-2-a-influencia-da-familia-e-dos-profissionais-de-saude/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>11 Outubro   2017</h5><h4><a href="http://noticias.up.pt/coliseuporto-recebe-algumas-das-melhores-tunas-do-pais/" rel="nofollow" class="title">Coliseu do Porto recebe algumas das melhores tunas do país</a></h4><p>A 19ª Edição do PortusCalle, organizada pela Tuna de Engenharia da Universidade do Porto, realiza-se dia 14 de outubro. Bilhetes disponíveis e à venda.  ... <a href="http://noticias.up.pt/coliseuporto-recebe-algumas-das-melhores-tunas-do-pais/" rel="nofollow">Ler toda a notícia</a></p></li>
+<li><h5>14 Outubro   2017</h5><h4><a href="http://noticias.up.pt/cintesis-da-inicio-a-estudo-piloto-para-colmatar-defice-de-iodo-nas-criancas/" rel="nofollow" class="title">CINTESIS dá início a estudo piloto para colmatar défice de iodo nas crianças</a></h4><p>Estudo com mais de 700 crianças entre os 6 e os 12 anos de idade tem por objetivo diminuir a carência de iodo na alimentação dos jovens em idade escolar.  ... <a href="http://noticias.up.pt/cintesis-da-inicio-a-estudo-piloto-para-colmatar-defice-de-iodo-nas-criancas/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>13 Outubro   2017</h5><h4><a href="http://noticias.up.pt/artur-santos-silva-e-o-novo-presidente-do-conselho-geral-da-u-porto/" rel="nofollow" class="title">Artur Santos Silva é o novo Presidente do Conselho Geral da U.Porto</a></h4><p>O gestor vai liderar durante os próximos quatros um dos principais órgãos de governo da Universidade, responsável - por exemplo - pela eleição do Reitor ou a fixação do valor das propinas. ... <a href="http://noticias.up.pt/artur-santos-silva-e-o-novo-presidente-do-conselho-geral-da-u-porto/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>13 Outubro   2017</h5><h4><a href="http://noticias.up.pt/u-porto-acolhe-novos-estudantes-da-segunda-e-terceira-fases/" rel="nofollow" class="title">U.Porto acolhe novos estudantes da segunda e terceira fases</a></h4><p>Novos estudantes podem levantar o seu kit nos dias 18 e 19 de outubro, entre as 16h00 e as 19h00, no E-Learning Café Asprela e E-Learning Café Jardim Botânico, respetivamente. ... <a href="http://noticias.up.pt/u-porto-acolhe-novos-estudantes-da-segunda-e-terceira-fases/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>13 Outubro   2017</h5><h4><a href="http://noticias.up.pt/fep-e-ordem-dos-economistas-promovem-2-a-edicao-da-economia-e-o-futuro/" rel="nofollow" class="title">FEP e Ordem dos Economistas promovem 2.ª edição da Economia e o Futuro</a></h4><p>Primeira conferência da segunda edição do ciclo ¿A Economia e o Futuro¿ realiza-se a 19 de outubro, às 18h00, no Salão Nobre da Faculdade de Economia da U.Porto ... <a href="http://noticias.up.pt/fep-e-ordem-dos-economistas-promovem-2-a-edicao-da-economia-e-o-futuro/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>13 Outubro   2017</h5><h4><a href="http://noticias.up.pt/feira-de-emprego-com-mais-de-3000-oportunidades-de-carreira/" rel="nofollow" class="title">Feira de Emprego da U.Porto com mais de 3000 oportunidades de carreira</a></h4><p>Mais de 3000 empregos e estágios vão estar disponíveis, de 17 a 18 de outubro, na Exponor. Estudantes da U.Porto com transporte gratuito entre as faculdades e a feira. ... <a href="http://noticias.up.pt/feira-de-emprego-com-mais-de-3000-oportunidades-de-carreira/" rel="nofollow">Ler toda a notícia</a></p></li><li><h5>13 Outubro   2017</h5><h4><a href="http://noticias.up.pt/dia-mundial-da-alimentacao-celebrado-pela-fcnaup/" rel="nofollow" class="title">Dia Mundial da Alimentação celebrado pela FCNAUP</a></h4><p>"Rotulagem Nutricional ¿ Os cidadãos e a informação sobre Nutrição¿ é o tema da sessão com que a FCNAUP vai assinalar o Dia Mundial da Alimentação, às 14h30 do dia 16 de outubro, na FCUP. ... <a href="http://noticias.up.pt/dia-mundial-da-alimentacao-celebrado-pela-fcnaup/" rel="nofollow">Ler toda a notícia</a></p></li>
 </ul>
 	</aside> -->
 <!-- end news column -->
