+++ scrapper/scrapper/items.py
@@ -12,3 +12,16 @@ class ScrapperItem(scrapy.Item):
     # define the fields for your item here like:
     # name = scrapy.Field()
     pass
+
+
+class FinalSchedule(scrapy.Item):
+    course = scrapy.Field()
+    date = scrapy.Field()
+    title = scrapy.Field()
+    text = scrapy.Field()
+    duration = scrapy.Field()
+    acronym = scrapy.Field()
+    professor = scrapy.Field()
+    prof_acro = scrapy.Field()
+    id_class = scrapy.Field()
+    location = scrapy.Field()
