+++ scrapper/scrapper/pipelines.py
@@ -4,8 +4,23 @@
 #
 # Don't forget to add your pipeline to the ITEM_PIPELINES setting
 # See: http://doc.scrapy.org/en/latest/topics/item-pipeline.html
+import json
 
 
 class ScrapperPipeline(object):
     def process_item(self, item, spider):
+        return item
+
+
+class JsonWriterPipeline(object):
+
+    def open_spider(self, spider):
+        self.file = open('clases.json', 'w')
+
+    def close_spider(self, spider):
+        self.file.close()
+
+    def process_item(self, item, spider):
+        line = json.dumps(dict(item)) + "\n"
+        self.file.write(line)
         return item
